(REF) Move PrimitiveFunctionVisitor
